requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.5
numpy
scikit-learn
joblib
pypdf
rapidfuzz
torch
torchvision
sentence-transformers
weaviate-client
